+++ src/perl/lib/WTSI/NPG/Genotyping/Fluidigm/Subscriber.pm
@@ -3,7 +3,7 @@ package WTSI::NPG::Genotyping::Fluidigm::Subscriber;
 
 use Moose;
 use JSON;
-use List::AllUtils qw(all natatime reduce uniq);
+use List::AllUtils qw(all reduce uniq);
 use Try::Tiny;
 
 use WTSI::NPG::Genotyping::Call;
@@ -19,10 +19,6 @@ our $VERSION = '';
 our $NO_CALL_GENOTYPE = 'NN';
 
 our $CHROMOSOME_JSON_ATTR = 'chromosome_json';
-
-# The largest number of bind variables iRODS supports for 'IN'
-# queries.
-our $BATCH_QUERY_CHUNK_SIZE = 100;
 
 with 'WTSI::DNAP::Utilities::Loggable', 'WTSI::NPG::Annotation',
   'WTSI::NPG::Genotyping::Annotation', 'WTSI::NPG::Genotyping::Subscription';
@@ -52,7 +48,7 @@ has '_plex_name_attr' =>
 
 sub get_assay_resultset {
     my ($self, $sample_identifier, @query_specs) = @_;
-    my $resultsets = $self->get_assay_resultsets
+    my ($resultsets, $vcf_meta) = $self->get_assay_resultsets_and_vcf_metadata
       ([$sample_identifier], @query_specs);
 
     my $num_samples = scalar keys %$resultsets;
@@ -72,86 +68,42 @@ sub get_assay_resultset {
     return shift @resultsets;
 }
 
-=head2 get_assay_resultsets
+=head2 get_assay_resultsets_and_vcf_metadata
 
   Arg [1]    : ArrayRef[Str] sample identifier (dcterms:identifier)
   Arg [n]    : Optional additional query specs as ArrayRefs.
 
   Example    : $sub->get_assay_resultsets(['0123456789'], [study => 12345]);
   Description: Fetch assay result sets by SNP set, sample and other optional
-               criteria.
-  Returntype : HashRef[ArrayRef[
-                 WTSI::NPG::Genotyping::Fluidigm::AssayResultSet]] indexed
+               criteria. Also finds associated VCF metadata.
+  Returntype : - HashRef[ArrayRef[
+                  WTSI::NPG::Genotyping::Fluidigm::AssayResultSet]] indexed
                by sample identifier. Each ArrayRef will usually contain one
                item, but may contain more if multiple assays match the
                search criteria.
+               - HashRef[ArrayRef[Str]] containing VCF metadata.
 
 =cut
 
-sub get_assay_resultsets {
+sub get_assay_resultsets_and_vcf_metadata {
   my ($self, $sample_identifiers, @query_specs) = @_;
 
-  defined $sample_identifiers or
-    $self->logconfess('A defined sample_identifiers argument is required');
-  ref $sample_identifiers eq 'ARRAY' or
-    $self->logconfess('The sample_identifiers argument must be an ArrayRef');
-  _are_unique($sample_identifiers) or
-    $self->logconfess('The sample_identifiers argument contained duplicate ',
-                      'values: [', join(q{, }, @$sample_identifiers), ']');
+  my @obj_paths =
+      $self->find_object_paths($sample_identifiers, @query_specs);
 
-  my $num_samples = scalar @$sample_identifiers;
-  my $chunk_size = $BATCH_QUERY_CHUNK_SIZE;
-
-  $self->debug("Getting results for $num_samples samples in chunks of ",
-               $chunk_size);
-
-  my @obj_paths;
-  my $iter = natatime $chunk_size, @$sample_identifiers;
-  while (my @ids = $iter->()) {
-    my @id_obj_paths = $self->irods->find_objects_by_meta
-      ($self->data_path,
-       [$self->_plex_name_attr => $self->snpset_name],
-       [$self->dcterms_identifier_attr => \@ids, 'in'], @query_specs);
-    push @obj_paths, @id_obj_paths;
-  }
-
-  my @resultsets = map {
-    WTSI::NPG::Genotyping::Fluidigm::AssayResultSet->new
-        (WTSI::NPG::Genotyping::Fluidigm::AssayDataObject->new
-         ($self->irods, $_));
+  # generate array of AssayDataObjects
+  # use to construct indexed AssayResultSets *and* find metadata
+  my @data_objects = map {
+      WTSI::NPG::Genotyping::Fluidigm::AssayDataObject->new
+            ($self->irods, $_);
   } @obj_paths;
-
-  # Index the results by sample identifier. The identifier is
-  # guaranteed to be present in the metadata because it was in the
-  # search criteria. No assumptions can be made about the order in
-  # which iRODS has returned the results, with respect to the
-  # arguments of the 'IN' clause.
-  my %resultsets_index;
-  foreach my $sample_identifier (@$sample_identifiers) {
-    unless (exists $resultsets_index{$sample_identifier}) {
-      $resultsets_index{$sample_identifier} = [];
-    }
-
-    my @sample_resultsets = grep {
-      $_->data_object->get_avu($self->dcterms_identifier_attr,
-                               $sample_identifier) } @resultsets;
-
-    $self->debug("Found ", scalar @sample_resultsets, " resultsets for ",
-                 "sample '$sample_identifier'");
-
-    # Sanity check that the contents are consistent
-    my @sample_names = map { $_->sample_name } @sample_resultsets;
-    unless (all { $_ eq $sample_names[0] } @sample_names) {
-      $self->logconfess("The resultsets found for sample AVU value ",
-                        "'$sample_identifier': did not all have the same ",
-                        "sample name in their data: [",
-                        join(q{, }, @sample_names), "]");
-    }
-
-    push @{$resultsets_index{$sample_identifier}}, @sample_resultsets;
-  }
-
-  return \%resultsets_index;
+  my @resultsets = map {
+      WTSI::NPG::Genotyping::Fluidigm::AssayResultSet->new($_);
+  } @data_objects;
+  my $resultsets_index =
+      $self->find_resultsets_index(\@resultsets, $sample_identifiers);
+  my $vcf_meta = $self->vcf_metadata_from_irods(\@data_objects);
+  return ($resultsets_index, $vcf_meta);
 }
 
 =head2 get_calls
@@ -185,7 +137,7 @@ sub get_calls {
   my @resultsets = @{$assay_resultsets};
   my $num_resultsets = scalar @resultsets;
 
-  my @sample_names = map { $_->sample_name } @resultsets;
+  my @sample_names = map { $_->canonical_sample_id } @resultsets;
   unless (all { $_ eq $sample_names[0] } @sample_names) {
     $self->logconfess('The assay_resultsets must belong to the same sample: [',
                       join(q{, }, @sample_names, ']'));